// src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      gigs: {
        Row: {
          budget: number | null
          created_at: string
          deadline: string | null
          description: string | null
          id: number
          location: string | null
          skills_required: string[] | null
          title: string
          type: Database["public"]["Enums"]["gig_type"] | null
          user_id: string
        }
        Insert: {
          budget?: number | null
          created_at?: string
          deadline?: string | null
          description?: string | null
          id?: never
          location?: string | null
          skills_required?: string[] | null
          title: string
          type?: Database["public"]["Enums"]["gig_type"] | null
          user_id: string
        }
        Update: {
          budget?: number | null
          created_at?: string
          deadline?: string | null
          description?: string | null
          id?: never
          location?: string | null
          skills_required?: string[] | null
          title?: string
          type?: Database["public"]["Enums"]["gig_type"] | null
          user_id?: string
        }
        Relationships: []
      }
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          bio: string | null
          collaboration_styles: string[] | null
          experience: string | null
          full_name: string | null
          genres: string[] | null
          github_url: string | null
          id: string
          instagram_url: string | null
          is_online: boolean | null
          is_verified: boolean | null
          level: number | null
          linkedin_url: string | null
          location: string | null
          portfolio: Json | null
          portfolio_url: string | null
          rating: number | null
          reviews: number | null
          role: string | null
          skills: string[] | null
          software_proficiencies: string[] | null
          specialization: string | null
          tags: string[] | null
          twitter_url: string | null
          updated_at: string | null
          username: string | null
          website: string | null
        }
        Insert: {
          avatar_url?: string | null
          bio?: string | null
          collaboration_styles?: string[] | null
          experience?: string | null
          full_name?: string | null
          genres?: string[] | null
          github_url?: string | null
          id: string
          instagram_url?: string | null
          is_online?: boolean | null
          is_verified?: boolean | null
          level?: number | null
          linkedin_url?: string | null
          location?: string | null
          portfolio?: Json | null
          portfolio_url?: string | null
          rating?: number | null
          reviews?: number | null
          role?: string | null
          skills?: string[] | null
          software_proficiencies?: string[] | null
          specialization?: string | null
          tags?: string[] | null
          twitter_url?: string | null
          updated_at?: string | null
          username?: string | null
          website?: string | null
        }
        Update: {
          avatar_url?: string | null
          bio?: string | null
          collaboration_styles?: string[] | null
          experience?: string | null
          full_name?: string | null
          genres?: string[] | null
          github_url?: string | null
          id?: string
          instagram_url?: string | null
          is_online?: boolean | null
          is_verified?: boolean | null
          level?: number | null
          linkedin_url?: string | null
          location?: string | null
          portfolio?: Json | null
          portfolio_url?: string | null
          rating?: number | null
          reviews?: number | null
          role?: string | null
          skills?: string[] | null
          software_proficiencies?: string[] | null
          specialization?: string | null
          tags?: string[] | null
          twitter_url?: string | null
          updated_at?: string | null
          username?: string | null
          website?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never